notas/dash/entre outros

+++ database/migrations/2018_07_26_231652_create_tb_services_table.php
@@ -16,7 +16,7 @@ class CreateTbServicesTable extends Migration
         Schema::create('tb_services', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name', 25);
-            $table->double('price', 10, 2);
+            $table->decimal('price', 10, 2);
             $table->timestamps();
         });
     }
